+++ newlib/libm/complex/cpowl.c
@@ -44,7 +44,7 @@ cpowl(long double complex a, long double complex z)
 	y = cimagl(z);
 	absa = cabsl(a);
 	if (absa == 0.0L) {
-		return (0.0L + 0.0L * (double complex) I);
+		return (0.0L + 0.0L * (long double complex) I);
 	}
 	arga = cargl(a);
 	r = powl(absa, x);
@@ -53,7 +53,7 @@ cpowl(long double complex a, long double complex z)
 		r = r * expl(-y * arga);
 		theta = theta + y * logl(absa);
 	}
-	w = r * cosl(theta) + (r * sinl(theta)) * (double complex) I;
+	w = r * cosl(theta) + (r * sinl(theta)) * (long double complex) I;
 	return w;
 }
