+++ newlib/libm/ld/ld128/k_sinl.c
@@ -51,7 +51,7 @@ __kernel_sinl(long double x, long double y, int iy)
 	z	=  x*x;
 	v	=  z*x;
 	r	=  S2+z*(S3+z*(S4+z*(S5+z*(S6+z*(S7+z*(S8+
-	    z*(S9+z*(S10+z*(S11+z*S12)))))))));
+	    z*((long double)S9+z*((long double)S10+z*((long double)S11+z*(long double)S12)))))))));
 	if(iy==0) return x+v*(S1+z*r);
-	else      return x-((z*(half*y-v*r)-y)-v*S1);
+	else      return x-((z*((long double)half*y-v*r)-y)-v*S1);
 }
