+++ newlib/libm/ld/ld80/s_nexttoward.c
@@ -32,7 +32,7 @@ nexttoward(__float64 x, long double y)
 
 	if(((ix>=0x7ff00000)&&((ix-0x7ff00000)|lx)!=0) ||   /* x is nan */
 	   ((iy>=0x7fff)&&(hy|ly)!=0))		/* y is nan */
-	   return x+y;
+	   return (long double)x+y;
 	if((long double) x==y) return y;	/* x=y, return y */
 	if((ix|lx)==0) {			/* x == 0 */
 	    volatile __float64 u;
